pyproject.toml
[project]
name = "tic-tac-toe-pro"
version = "0.1.0"
description = "A real-time multiplayer Tic-Tac-Toe game built with FastAPI and Redis."
authors = ["Shubhajit Halder <shubhajit07@yahoo.com>"]
requires-python = ">=3.10"

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "redis",
    "python-dotenv"
]

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "pytest"
]
